opgesplitst in bestanden

+++ src/main.py
@@ -29,7 +29,7 @@ if __name__ == "__main__":
         SubConfigurationMap.get_configuration(SimulationConfiguration, "standard"),
         SubConfigurationMap.get_configuration(CPGConfiguration, "standard"),
         SubConfigurationMap.get_configuration(RandomConfiguration, "standard"),
-        SubConfigurationMap.get_configuration(GeneticConfiguration, "short"),
+        SubConfigurationMap.get_configuration(GeneticConfiguration, "long"),
         SubConfigurationMap.get_configuration(ControllerConfiguration, "standard"),
     )
